upgrade dependencies + remove deprecation warnings + fix Day10, 2020 + update tests

+++ src/test/scala/org/base/advent/_2020/Day09Spec.scala
@@ -7,7 +7,7 @@ import org.mockito.scalatest.IdiomaticMockito
 object Day09Spec extends IdiomaticMockito {
   trait Environment {
     val day = new Day09
-    val example1: Seq[Int] =
+    val example1: Seq[Long] =
       Seq(35, 20, 15, 25, 47, 40, 62, 55, 65, 95, 102, 117, 150, 182, 127, 219, 299, 277, 309, 576)
   }
 }
